% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_neighborhood.R
\name{add_neighborhood}
\alias{add_neighborhood}
\title{Add neighborhood based on census tract}
\usage{
add_neighborhood(.x, vintage = c("2010", "2020"))
}
\arguments{
\item{.x}{a data frame containing a census tract column named
\code{census_tract_id}, \code{census_tract_id_2010}, or \code{census_tract_id_2020}}

\item{vintage}{a string specifying to use the \code{2010} or \code{2020} census tract to neighborhood lookup table;
if set will override any vintage found in census tract id column names}
}
\description{
This function uses the internal data objects, \code{hamilton_tract_to_cincy_neighborhood_2010} and
\code{hamilton_tract_to_cincy_neighborhood_2020}, to add a neighborhood
from \code{cincy::neigh_cchmc_2010} and \code{cincy::neigh_cchmc_2020} for each
row of a data frame.
}
\details{
The vintage of tracts is automatically inferred based on the
name of the found tract column (\code{census_tract_id_2010} or \code{census_tract_id_2020}), but
can be specified using the vintage argument. If just the column \code{census_tract_id}
exists, then a default of 2010 will be used.
}
\examples{
add_neighborhood(tract_tigris_2010)
}
