% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blockgroup_tigris_2000}
\alias{blockgroup_tigris_2000}
\alias{blockgroup_tigris_2010}
\alias{blockgroup_tigris_2020}
\title{Block Groups}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 736 rows and 2 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 697 rows and 2 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 678 rows and 2 columns.
}
\source{
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.2000.html}{2000 TIGER/Line Shapefiles}

\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.2010.html}{2010 TIGER/Line Shapefiles}

\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.2020.html}{2020 TIGER/Line Shapefiles}
Block Groups (2000 TIGER/Line)
}
\usage{
blockgroup_tigris_2000

blockgroup_tigris_2010

blockgroup_tigris_2020
}
\description{
Census block groups for Hamilton County from 2000, 2010, and 2020.
}
\details{
Block groups were downloaded directly from the Census Bureau using the \href{https://github.com/walkerke/tigris}{tigris} package
}
\keyword{datasets}
