% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_swoh_2010}
\alias{county_swoh_2010}
\alias{county_hlthv_2010}
\alias{county_hlthvts_2010}
\alias{county_7cc_2010}
\alias{county_8cc_2010}
\title{Counties}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 4 rows and 6 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 8 rows and 6 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 15 rows and 6 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 7 rows and 6 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 8 rows and 6 columns.
}
\source{
see data-raw/data.R for county FIPS listed for each grouping
Southwest Ohio
}
\usage{
county_swoh_2010

county_hlthv_2010

county_hlthvts_2010

county_7cc_2010

county_8cc_2010
}
\description{
County groupings are derived from CCHMC operational definitions
}
\keyword{datasets}
