% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dep_index}
\alias{dep_index}
\title{Deprivation Index}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 222 rows and 9 columns.
}
\source{
The \href{https://geomarker.io/dep_index/2018_dep_index/}{2018 deprivation_index}. See \code{data-raw/make_dep_index_data.R}.
}
\usage{
dep_index
}
\description{
An example census tract-level dataset.  A simple features object of all census tracts in
Hamilton County, OH with values for the 2018 deprivation index and its six components.
}
\keyword{datasets}
