% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{districts_tigris_2011}
\alias{districts_tigris_2011}
\alias{districts_tigris_2013}
\title{Congressional Districts (2011 TIGER/Line)}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 2 rows and 2 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 2 rows and 2 columns.
}
\usage{
districts_tigris_2011

districts_tigris_2013
}
\description{
Congressional Districts (2011 TIGER/Line)

Congressional Districts (2013 TIGER/Line)
}
\keyword{datasets}
