% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hamilton_tract_to_cincy_neighborhood_2010}
\alias{hamilton_tract_to_cincy_neighborhood_2010}
\title{Tract to Neighborhood Lookup 2010}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 222 rows and 2 columns.
}
\usage{
hamilton_tract_to_cincy_neighborhood_2010
}
\description{
Tract to Neighborhood Lookup 2010
}
\keyword{datasets}
