% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hamilton_tract_to_cincy_neighborhood_2020}
\alias{hamilton_tract_to_cincy_neighborhood_2020}
\title{Tract to Neighborhood Lookup 2020}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 226 rows and 2 columns.
}
\usage{
hamilton_tract_to_cincy_neighborhood_2020
}
\description{
Tract to Neighborhood Lookup 2020
}
\keyword{datasets}
