% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neigh_ccc}
\alias{neigh_ccc}
\title{Neighborhood: Cincinnati Community Council (CCC)}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 75 rows and 2 columns.
}
\source{
\href{http://cagismaps.hamilton-co.org/cagisportal}{CAGIS}
\href{http://cagis.org/Opendata/?}{Open Data} Community Council (CCC) Neighborhoods
}
\usage{
neigh_ccc
}
\description{
Neighborhood: Cincinnati Community Council (CCC)
}
\keyword{datasets}
