% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neigh_cchmc_2020}
\alias{neigh_cchmc_2020}
\title{Neighborhood: CCHMC 2020}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 81 rows and 2 columns.
}
\source{
See \code{data-raw/make_tract_neigh_crosswalk_data.R} for tract to neighborhood lookup table
}
\usage{
neigh_cchmc_2020
}
\description{
Neighborhood: CCHMC 2020
}
\keyword{datasets}
