% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neigh_sna}
\alias{neigh_sna}
\title{Neighborhood: Statistical Neighborhood Approximations (SNA)}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 50 rows and 2 columns.
}
\source{
\href{http://cagismaps.hamilton-co.org/cagisportal}{CAGIS}
\href{http://cagis.org/Opendata/?}{Open Data} Community Council (CCC) Neighborhoods
}
\usage{
neigh_sna
}
\description{
Neighborhood: Statistical Neighborhood Approximations (SNA)
}
\keyword{datasets}
