% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neigh_ccc}
\alias{neigh_ccc}
\alias{neigh_sna}
\title{Neighborhoods}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 75 rows and 2 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 50 rows and 2 columns.
}
\usage{
neigh_ccc

neigh_sna
}
\description{
Neighborhoods for Hamilton County defined in \href{http://cagismaps.hamilton-co.org/cagisportal}{CAGIS} \href{http://cagis.org/Opendata/?}{Open Data}
Neighborhoods (CCC)
}
\keyword{datasets}
