% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zcta_tigris_2000}
\alias{zcta_tigris_2000}
\alias{zcta_tigris_2010}
\alias{zcta_tigris_2020}
\title{ZCTAs}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 53 rows and 2 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 54 rows and 2 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 55 rows and 2 columns.
}
\source{
ZCTAs were downloaded directly from the Census Bureau using the \href{https://github.com/walkerke/tigris}{tigris} package
ZCTA (2000 TIGER/Line)
}
\usage{
zcta_tigris_2000

zcta_tigris_2010

zcta_tigris_2020
}
\description{
\href{https://www.census.gov/programs-surveys/geography/guidance/geo-areas/zctas.html}{ZIP Code Tabulation Areas (ZCTAs)}
for Ohio that (at least partially) intersect with Hamilton County from 2000, 2010, and 2020.
}
\keyword{datasets}
