% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codec_counties}
\alias{codec_counties}
\title{CODEC Counties}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 15 rows and 4 columns.
}
\usage{
codec_counties
}
\description{
CODEC Counties
}
\keyword{datasets}
