% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_7cc_2010}
\alias{county_7cc_2010}
\title{Counties: 7 County Region}
\format{
A simple features data frame with 7 rows and 6 columns:
\describe{
\item{county_name}{County Name}
\item{county_id}{County ID}
\item{state_name}{State Name}
\item{state_id}{State ID}
\item{geoid}{GEOID (state_id + county_id)}
\item{geometry}{simple features geometry column}
}
}
\source{
see data-raw/data.R for county FIPS listed for each grouping
}
\usage{
county_7cc_2010
}
\description{
County groupings are derived from CCHMC operational definitions
}
\examples{
tmap::tm_shape(county_7cc_2010) +
  tmap::tm_polygons(col = "county_name",
                    legend.show = FALSE) +
  tmap::tm_text(text = "county_name")
}
\keyword{datasets}
