% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_hlthv_2010}
\alias{county_hlthv_2010}
\title{Counties: Healthvine}
\format{
A simple features data frame with 15 rows and 6 columns:
\describe{
\item{county_name}{County Name}
\item{county_id}{County ID}
\item{state_name}
\item{state_id}
\item{geoid}{GEOID (state_id + county_id)}
\item{geometry} simple features geometry column
}
}
\source{
see data-raw/data.R for county FIPS listed for each grouping
}
\usage{
county_hlthv_2010
}
\description{
County groupings are derived from CCHMC operational definitions
}
\examples{
plot(county_hlthv_2010["county_name"], key.pos = 1)
}
\keyword{datasets}
