% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate}
\alias{interpolate}
\title{Spatially interpolate community-level data}
\usage{
interpolate(from, to, weights = c("pop", "homes", "area"))
}
\arguments{
\item{from}{sf object with a neighborhood, census tract, or zcta column and
numeric values to be interpolated into target geographies}

\item{to}{sf object of target geography (\strong{must} be one of the cincy:: geography objects)}

\item{weights}{use one of "pop" (population), "homes", or "area" from the
2020 census block estimates to interpolate the values}
}
\description{
Weights at the census block-level are used to spatially interpolate different geographies.
Block-level total population, total number of homes, or total area from the 2020 Census
can be chosen to use for the weights.
All \emph{numeric} variables in \code{from} are interpolated \emph{non-extensively},
except for any numeric variables that start with \code{n_}, which are interpolated
\emph{extensively}.
}
\examples{
# interpolate 2018 deprivation index to ZIP code level
interpolate(dep_index, cincy::zcta_tigris_2010, "pop")
# interpolate 2018 deprivation index to  2020 census tracts
interpolate(dep_index, cincy::tract_tigris_2020, "pop")
}
