% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate}
\alias{interpolate}
\title{Spatially interpolate community-level data}
\usage{
interpolate(from, to, weights = c("pop", "homes", "area"))
}
\arguments{
\item{from}{sf object with a neighborhood, census tract, or zcta column and
numeric values to be interpolated into target geographies}

\item{to}{sf object of target geography (\strong{must} be one of the cincy:: geography objects)}

\item{weights}{use one of "pop" (population), "homes", or "area" from the
2020 census block estimates to interpolate the values}
}
\description{
This function will interpolate only \emph{numeric} variables with names starting with
\code{n_}, \code{median_}, or \code{fraction_}.  Internally, \code{fraction_} and \code{median_} variables
are interpolated \emph{non-extensively}, but \code{n_} variables are interpolated \emph{extensively}.
Weights at the census block level are used to spatially interpolate different geographies.
Block-level total population, total number of homes, or total area from the 2020 Census
can be chosen to use for the weights.
}
\examples{
# interpolate 2018 deprivation index to ZIP code level
interpolate(dep_index, cincy::zcta_tigris_2010, "pop")
# interpolate 2018 deprivation index to  2020 census tracts
interpolate(dep_index, cincy::tract_tigris_2020, "pop")
}
