% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neigh_cchmc}
\alias{neigh_cchmc}
\title{Neighborhood: CCHMC}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 81 rows and 2 columns.
}
\source{
See \code{data-raw/hamilton_tract_to_cincy_neighborhood.csv} for tract to neighborhood lookup table
}
\usage{
neigh_cchmc
}
\description{
Neighborhood: CCHMC
}
\keyword{datasets}
