% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tract_tigris_2000}
\alias{tract_tigris_2000}
\alias{tract_tigris_2010}
\alias{tract_tigris_2020}
\title{Tracts}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 230 rows and 2 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 222 rows and 2 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 226 rows and 2 columns.
}
\source{
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.2000.html}{TIGER/Line Shapefile}

\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.2010.html}{TIER/Line Shapefile}

\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.2020.html}{TIGER/Line Shapefile}
}
\usage{
tract_tigris_2000

tract_tigris_2010

tract_tigris_2020
}
\description{
Census Tracts for Hamilton County from 2000, 2010, and 2020.
}
\details{
Tracts were downloaded directly from the Census Bureau using the \href{https://github.com/walkerke/tigris}{tigris} package
Census Tracts (2000 TIGER/Line)
}
\keyword{datasets}
